Actually connect to DB when checking connections
When checking the DB connection (e.g. by requesting `POST /api/reports?client=simple_report&option=CheckConnections`), the check previously succeeded if a connection had been made to the database before, even if that connection had since died. This ensures we actually transact with the database to make sure the connection is still live when it is checked.